+++ dsmr-graphql/src/main/java/nl/basjes/dsmr/graphql/GraphQlConfig.java
@@ -1,0 +1,53 @@
+/*
+ * Dutch Smart Meter Requirements (DSMR) Toolkit
+ * Copyright (C) 2019-2021 Niels Basjes
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package nl.basjes.dsmr.graphql;
+
+import graphql.scalars.ExtendedScalars;
+import nl.basjes.dsmr.DSMRTelegram;
+import nl.basjes.dsmr.graphql.scalars.DurationScalar;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.graphql.execution.RuntimeWiringConfigurer;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Sinks;
+
+@Configuration
+public class GraphQlConfig {
+    @Bean
+    public RuntimeWiringConfigurer runtimeWiringConfigurer() {
+        return wiringBuilder -> wiringBuilder
+                .scalar(ExtendedScalars.GraphQLLong)
+                .scalar(ExtendedScalars.DateTime)
+                .scalar(DurationScalar.INSTANCE);
+    }
+
+    @Bean
+    public Sinks.Many<DSMRTelegram> sink() {
+        return Sinks.many()
+            .multicast()
+            .directBestEffort();
+//        .onBackpressureBuffer(Queues.SMALL_BUFFER_SIZE, false);
+    }
+
+    @Bean
+    public Flux<DSMRTelegram> flux() {
+        return sink().asFlux();
+    }
+
+}